+++ tests/crescent/commands/test_command_function.py
@@ -7,7 +7,6 @@ from crescent.internal.app_command import AppCommand, AppCommandMeta
 
 
 def test_command_function():
-
     @command(description="1234")
     async def callback(
         ctx: Context,
@@ -26,25 +25,19 @@ def test_command_function():
             description="1234",
             options=[
                 CommandOption(
-                    type=OptionType.STRING,
-                    name="arg_1",
-                    description="1234",
-                    is_required=True,
+                    type=OptionType.STRING, name="arg_1", description="1234", is_required=True
                 ),
                 CommandOption(
-                    type=OptionType.STRING,
-                    name="arg_2",
-                    description="1234",
-                    is_required=False,
+                    type=OptionType.STRING, name="arg_2", description="1234", is_required=False
                 ),
                 CommandOption(
                     type=OptionType.FLOAT,
                     name="arg_3",
                     description="No Description",
                     is_required=False,
-                )
+                ),
             ],
-        ),
+        )
     )
 
     @command(name="test", description="1234", guild=12345678)
@@ -58,7 +51,7 @@ def test_command_function():
             guild_id=12345678,
             default_permission=UNDEFINED,
             description="1234",
-        ),
+        )
     )
 
     @message_command
@@ -67,11 +60,8 @@ def test_command_function():
 
     assert callback.metadata == AppCommandMeta(
         app=AppCommand(
-            type=CommandType.MESSAGE,
-            name="callback",
-            default_permission=UNDEFINED,
-            guild_id=None,
-        ),
+            type=CommandType.MESSAGE, name="callback", default_permission=UNDEFINED, guild_id=None
+        )
     )
 
     @user_command
@@ -80,9 +70,6 @@ def test_command_function():
 
     assert callback.metadata == AppCommandMeta(
         app=AppCommand(
-            type=CommandType.USER,
-            name="callback",
-            default_permission=UNDEFINED,
-            guild_id=None,
-        ),
+            type=CommandType.USER, name="callback", default_permission=UNDEFINED, guild_id=None
+        )
     )
