+++ tests/test_bot/test_bot.py
