+++ tests/test_bot/test_plugin.py
