+++ tests/utils/__init__.py
@@ -2,6 +2,4 @@ from typing import Sequence
 
 from tests.utils.mock_bot import MockBot
 
-__all__: Sequence[str] = (
-    "MockBot",
-)
+__all__: Sequence[str] = ("MockBot",)
