+++ tests/utils/mock_bot.py
@@ -1,9 +1,11 @@
 from typing import Any
+
+from crescent import Bot
 from crescent.internal.meta_struct import MetaStruct
 from crescent.internal.registry import CommandHandler
 
 
-class MockBot:
+class MockBot(Bot):
     def __init__(self) -> None:
 
         self.default_guild = None
@@ -13,6 +15,6 @@ class MockBot:
 
         self._command_handler = CommandHandler(self, [])
 
-    def include(self, command: MetaStruct[Any, Any]):
+    def include(self, command: MetaStruct[Any, Any]):  # type: ignore
         self._command_handler.register(command)
         return command
